+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 1.0.0
  December 14, 2021
 - initial commit and release
+
+## 1.1.0
+ December 17, 2021
+- utils/hyperlink.dart';
+- utils/iframe_widget.dart';
+- utils/web_utils.dart';